Berkat, fix trouble seed database

+++ database/migrations/2025_06_03_081618_create_ajuan_table.php
@@ -1,5 +1,3 @@
-// database/migrations/2025_06_17_000001_create_ajuan_table.php
-
 <?php
 
 use Illuminate\Database\Migrations\Migration;
@@ -14,10 +12,13 @@ return new class extends Migration
         Schema::create('ajuan', function (Blueprint $table) {
             $table->id('id_ajuan');
             $table->date('tanggal_ajuan')->default(DB::raw('CURRENT_TIMESTAMP'));
-            $table->unsignedInteger('id_admin')->nullable();
+            $table->unsignedBigInteger('id_admin')->nullable(); // Changed to unsignedBigInteger
             $table->timestamps();
             
-            $table->foreign('id_admin')->references('id_admin')->on('admin')->onDelete('cascade');
+            $table->foreign('id_admin')
+                  ->references('id_admin')
+                  ->on('admin')
+                  ->onDelete('cascade');
         });
     }
