+++ database/migrations/2025_06_17_081245_create_pemohon_bidang_table.php
@@ -1,5 +1,3 @@
-// database/migrations/2025_06_17_000008_create_pemohon_bidang_table.php
-
 <?php
 
 use Illuminate\Database\Migrations\Migration;
@@ -11,12 +9,18 @@ return new class extends Migration
     public function up()
     {
         Schema::create('pemohon_bidang', function (Blueprint $table) {
-            $table->unsignedInteger('id_bidang');
-            $table->unsignedInteger('id_pemohon');
+            $table->unsignedBigInteger('id_bidang');
+            $table->unsignedBigInteger('id_pemohon');
             
             $table->primary(['id_bidang', 'id_pemohon']);
-            $table->foreign('id_bidang')->references('id_bidang')->on('bidang_kerjasama')->onDelete('cascade');
-            $table->foreign('id_pemohon')->references('id_pemohon')->on('pemohon')->onDelete('cascade');
+            $table->foreign('id_bidang')
+                  ->references('id_bidang')
+                  ->on('bidang_kerjasama')
+                  ->onDelete('cascade');
+            $table->foreign('id_pemohon')
+                  ->references('id_pemohon')
+                  ->on('pemohon')
+                  ->onDelete('cascade');
         });
     }
